refactor into its own class
